Flesh out the endpoint for uploading replays.
This will eventually be called by a watcher who is querying hotsapi to
see if there are unprocessed uploads, but will support generic uploading
of replays.

It does a quick validation pass to see if the file looks a real replay
files, and will generate a UUID for it.  We can then track it in the
database, make sure we aren't storing duplicated matches.

After that happens, we pass it off to a processor for full parsing and
event extraction.  This part is stubbed out right now, and doesn't do
anything yet.  We track the status of a replay -- whether it's been
processed or not -- so we can have some sort of management command to
force a reprocess of files which are stale and haven't yet been
processed, possibly due a crash or bug that drops the message on the
floor.